requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.2
openpyxl==3.1.5
jinja2==3.1.4
playwright==1.40.0
sqlalchemy==2.0.23
alembic==1.13.1
pytest-asyncio==0.21.1
pytest==7.4.3
requests==2.31.0
python-multipart
PyJWT==2.10.1
pydantic-settings==2.11.0
